+++ src/main/java/com/sanedge/ecommerce/models/category/CategoriesMonthPrice.java
@@ -9,9 +9,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CategoriesMonthPrice {
     private String month;
-    private Integer categoryId;
+    private Long categoryId;
     private String categoryName;
-    private Integer orderCount;
-    private Integer itemsSold;
+    private Long orderCount;
+    private Long itemsSold;
     private Long totalRevenue;
 }
